Style hero section
The styling for the hero section was done. Slight modifications were made to the HTML file (addition of more divs) to accomplish this.

A new directory for images was created and the image used in the hero section was added to it. The image credits were added to the README.md file