Style information section
CSS styling for the information section was done. Relative sizing units have been employed to improve responsiveness and to start getting used to using such units, with the exception of some images.

The README.md file was edited to include credits for images used in the information section (i.e. section-2)